+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## 2.9.36 - 2017-09-26
+### Changed
 - improved database connection setup
 - improved item export by using parent-child-relations, when possible
 
@@ -233,3 +236,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - export categories, products and reviews to Shopgate
 
 [Unreleased]: https://github.com/shopgate/interface-opencart/compare/2.9.35...HEAD
+[2.9.36]: